feat: Refactoring BookingSeatList Codes
BookingSeatList 컴포넌트에 있는 함수들 Util 파일로 분리
